+++ src/panel/main.ts
@@ -19,6 +19,7 @@ import {
 } from './pokemon-collection';
 import { PokemonElementState, PokemonPanelState } from './states';
 import { getRandomPokemonConfig } from '../common/pokemon-data';
+import { reactionController } from './reaction-controller';
 
 /* This is how the VS Code API can be invoked from the panel */
 declare global {
@@ -114,6 +115,7 @@ function startAnimations(
       });
     });
     pokemon.nextFrame();
+    reactionController.tick(pokemon, allPokemon);
     saveState(stateApi);
   }, 100);
 }
@@ -351,6 +353,7 @@ function removePokemonFromPanel(
 
   // Remove from collection immediately so rapid deletes of Pokemon don't interfere with each other
   allPokemon.removeFromCollection(message.name);
+  reactionController.forget(message.name);
   pokemon.collision.remove();
   pokemon.speech.remove();
   pokemonCounter = normalizePokemonCounter(pokemonCounter - 1);
@@ -658,6 +661,9 @@ export function pokemonPanelApp(
         break;
       case 'evolve-pokemon':
         evolvePokemonInPanel(message, basePokemonUri, stateApi);
+        break;
+      case 'pokemon-reaction':
+        reactionController.handleMessage(message, allPokemon);
         break;
       case 'reset-pokemon':
         var pokemonToRemove = [...allPokemon.pokemonCollection];